+++ packages/terra-site/src/examples/content-container/Index.jsx
@@ -1,27 +1,42 @@
-/* eslint-disable import/no-extraneous-dependencies */
+/* eslint-disable import/no-extraneous-dependencies, import/no-webpack-loader-syntax, import/first, import/no-unresolved, import/extensions */
 import React from 'react';
-import PropsTable from 'terra-props-table';
-import Markdown from 'terra-markdown';
+import DocTemplate from 'terra-doc-template';
 import ReadMe from 'terra-content-container/docs/README.md';
-import { version } from 'terra-content-container/package.json';
+import { name } from 'terra-content-container/package.json';
 
 // Component Source
-// eslint-disable-next-line import/no-webpack-loader-syntax, import/first, import/no-unresolved, import/extensions
 import ContentContainerSrc from '!raw-loader!terra-content-container/src/ContentContainer.jsx';
+
 // Example Files
 import ContentContainerStandard from './ContentContainerStandard';
+import ContentContainerStandardSrc from '!raw-loader!./ContentContainerStandard';
 import ContentContainerFill from './ContentContainerFill';
+import ContentContainerFillSrc from '!raw-loader!./ContentContainerFill';
 
-const ContentContainerExamples = () => (
-  <div>
-    <div id="version">Version: {version}</div>
-    <Markdown id="readme" src={ReadMe} />
-    <PropsTable id="props" src={ContentContainerSrc} />
-    <h2 id="standard">Standard Container</h2>
-    <ContentContainerStandard />
-    <h2 id="outline-variant">Fill Container</h2>
-    <ContentContainerFill />
-  </div>
+const DocPage = () => (
+  <DocTemplate
+    packageName={name}
+    readme={ReadMe}
+    srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
+    examples={[
+      {
+        title: 'Standard Container',
+        example: <ContentContainerStandard />,
+        source: ContentContainerStandardSrc,
+      },
+      {
+        title: 'Fill Container',
+        example: <ContentContainerFill />,
+        source: ContentContainerFillSrc,
+      },
+    ]}
+    propsTables={[
+      {
+        componentName: 'Content Container',
+        componentSrc: ContentContainerSrc,
+      },
+    ]}
+  />
 );
 
-export default ContentContainerExamples;
+export default DocPage;