+++ packages/terra-site/src/examples/heading/Index.jsx
@@ -1,6 +1,6 @@
-/* eslint-disable import/no-extraneous-dependencies */
+/* eslint-disable import/no-extraneous-dependencies, import/no-webpack-loader-syntax, import/first, import/no-unresolved, import/extensions */
 import React from 'react';
-import PropsTable from 'terra-props-table';
+import DocTemplate from 'terra-doc-template';
 import Markdown from 'terra-markdown';
 import ReadMe from 'terra-heading/docs/README.md';
 import HeadingLevelDocs from 'terra-heading/docs/heading-levels.md';
@@ -9,48 +9,69 @@ import HeadingColorDocs from 'terra-heading/docs/heading-colors.md';
 import HeadingWeightDocs from 'terra-heading/docs/heading-weights.md';
 import HeadingVisuallyHiddenDocs from 'terra-heading/docs/heading-visually-hidden.md';
 import HeadingVariationDocs from 'terra-heading/docs/heading-variations.md';
-import { version } from 'terra-heading/package.json';
+import { name } from 'terra-heading/package.json';
 
 // Component Source
-// eslint-disable-next-line import/no-webpack-loader-syntax, import/first, import/no-unresolved, import/extensions
 import HeadingSrc from '!raw-loader!terra-heading/src/Heading';
 
 // Example Files
 import HeadingColors from './HeadingColors';
+import HeadingColorsSrc from '!raw-loader!./HeadingColors';
 import HeadingLevels from './HeadingLevels';
+import HeadingLevelsSrc from '!raw-loader!./HeadingLevels';
 import HeadingSizes from './HeadingSizes';
+import HeadingSizesSrc from '!raw-loader!./HeadingSizes';
 import HeadingVisuallyHidden from './HeadingVisuallyHidden';
+import HeadingVisuallyHiddenSrc from '!raw-loader!./HeadingVisuallyHidden';
 import HeadingWeights from './HeadingWeights';
+import HeadingWeightsSrc from '!raw-loader!./HeadingWeights';
 import HeadingVariations from './HeadingVariations';
+import HeadingVariationsSrc from '!raw-loader!./HeadingVariations';
 
-const HeadingExamples = () => (
-  <div>
-    <div id="version">Version: {version}</div>
-    <Markdown id="readme" src={ReadMe} />
-    <PropsTable id="props" src={HeadingSrc} />
-    <Markdown id="heading-level-docs" src={HeadingLevelDocs} />
-    <HeadingLevels />
-    <br />
-    <br />
-    <Markdown id="heading-size-docs" src={HeadingSizeDocs} />
-    <HeadingSizes />
-    <br />
-    <br />
-    <Markdown id="heading-color-docs" src={HeadingColorDocs} />
-    <HeadingColors />
-    <br />
-    <br />
-    <Markdown id="heading-weight-docs" src={HeadingWeightDocs} />
-    <HeadingWeights />
-    <br />
-    <br />
-    <Markdown id="heading-visually-hidden-docs" src={HeadingVisuallyHiddenDocs} />
-    <HeadingVisuallyHidden />
-    <br />
-    <br />
-    <Markdown id="heading-variation-docs" src={HeadingVariationDocs} />
-    <HeadingVariations />
-  </div>
+const DocPage = () => (
+  <DocTemplate
+    packageName={name}
+    readme={ReadMe}
+    srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
+    examples={[
+      {
+        description: <Markdown id="heading-level-docs" src={HeadingLevelDocs} />,
+        example: <HeadingLevels />,
+        source: HeadingLevelsSrc,
+      },
+      {
+        description: <Markdown id="heading-size-docs" src={HeadingSizeDocs} />,
+        example: <HeadingSizes />,
+        source: HeadingSizesSrc,
+      },
+      {
+        description: <Markdown id="heading-color-docs" src={HeadingColorDocs} />,
+        example: <HeadingColors />,
+        source: HeadingColorsSrc,
+      },
+      {
+        description: <Markdown id="heading-weights-docs" src={HeadingWeightDocs} />,
+        example: <HeadingWeights />,
+        source: HeadingWeightsSrc,
+      },
+      {
+        description: <Markdown id="heading-visually-hidden-docs" src={HeadingVisuallyHiddenDocs} />,
+        example: <HeadingVisuallyHidden />,
+        source: HeadingVisuallyHiddenSrc,
+      },
+      {
+        description: <Markdown id="heading-variation-docs" src={HeadingVariationDocs} />,
+        example: <HeadingVariations />,
+        source: HeadingVariationsSrc,
+      },
+    ]}
+    propsTables={[
+      {
+        componentName: 'Heading',
+        componentSrc: HeadingSrc,
+      },
+    ]}
+  />
 );
 
-export default HeadingExamples;
+export default DocPage;