+++ packages/terra-site/src/examples/slide-group/Index.jsx
@@ -1,30 +1,40 @@
-/* eslint-disable import/no-extraneous-dependencies */
+/* eslint-disable import/no-extraneous-dependencies, import/no-webpack-loader-syntax, import/first, import/no-unresolved, import/extensions */
 import React from 'react';
-import PropsTable from 'terra-props-table';
-import Markdown from 'terra-markdown';
+import DocTemplate from 'terra-doc-template';
 import ReadMe from 'terra-slide-group/docs/README.md';
-import { version } from 'terra-slide-group/package.json';
+import { name } from 'terra-slide-group/package.json';
 
 // Component Source
-// eslint-disable-next-line import/no-webpack-loader-syntax, import/first, import/no-unresolved, import/extensions
 import SlideGroupSrc from '!raw-loader!terra-slide-group/src/SlideGroup.jsx';
 
 // Example Files
 import SlideGroupDemo from './SlideGroupDemo';
+import SlideGroupDemoSrc from '!raw-loader!./SlideGroupDemo';
 
-const SlideGroupExamples = () => (
-  <div>
-    <div id="version">Version: {version}</div>
-    <Markdown id="readme" src={ReadMe} />
-    <PropsTable id="props-slideGroup" src={SlideGroupSrc} />
-    <h2 id="slideGroup">Non-Animated SlideGroup</h2>
-    <br />
-    <SlideGroupDemo />
-    <br />
-    <h2 id="slideGroup-animated">Animated SlideGroup</h2>
-    <br />
-    <SlideGroupDemo isAnimated />
-  </div>
+const DocPage = () => (
+  <DocTemplate
+    packageName={name}
+    readme={ReadMe}
+    srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
+    examples={[
+      {
+        title: 'Non-Animated SlideGroup',
+        example: <SlideGroupDemo />,
+        source: SlideGroupDemoSrc,
+      },
+      {
+        title: 'Animated SlideGroup',
+        example: <SlideGroupDemo isAnimated />,
+        source: SlideGroupDemoSrc,
+      },
+    ]}
+    propsTables={[
+      {
+        componentName: 'Slide Group',
+        componentSrc: SlideGroupSrc,
+      },
+    ]}
+  />
 );
 
-export default SlideGroupExamples;
+export default DocPage;